string_utils: handle overflow correct in parse_byte_size_string()
This takes the overflow handling code from the kernel.

Link: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=32549

+++ src/lxc/string_utils.h
@@ -81,7 +81,7 @@ __hidden extern int lxc_safe_uint64(const char *numstr, uint64_t *converted, int
 __hidden extern int lxc_safe_int64_residual(const char *numstr, int64_t *converted, int base,
 					    char *residual, size_t residual_len);
 /* Handles B, kb, MB, GB. Detects overflows and reports -ERANGE. */
-__hidden extern int parse_byte_size_string(const char *s, int64_t *converted);
+__hidden extern int parse_byte_size_string(const char *s, long long int *converted);
 
 /*
  * Concatenate all passed-in strings into one path. Do not fail. If any piece
